Diego

+++ app/src/main/java/es/unir/cuentameuncuento/impls/BookDAOImpl.java
@@ -66,12 +66,13 @@ public class BookDAOImpl  {
                             if(task.isSuccessful()){
                                 Log.d("BookDAOImpl", "Callback valid Story=" + book.toString());
 
-                                storageImpl.create(SessionManager.currentStory.getIcon(), uniqueStoryImageUUID);
+                                storageImpl.create(SessionManager.currentStory.getIcon(), uniqueStoryImageUUID, callback);
 
-                                callback.onComplete(true, "Libro creado");
+//                                callback.onComplete(true, "Libro creado");
                             } else{
                                 Log.d("BookDAOImpl", task.getResult().toString());
                                 Log.d("BookDAOImpl", "Callback null Story=" + book.toString());
+
                                 callback.onComplete(false, "Operación fallida");
                             }
                         }
@@ -118,7 +119,8 @@ public class BookDAOImpl  {
                 if (task.isSuccessful()) {
                     //controller.refresh();
 
-                    storageImpl.delete(story.getIconID());
+                    if(!story.getIconID().isEmpty())
+                        storageImpl.delete(story.getIconID());
 
                     callback.onComplete(true, "Libro borrado");
                 } else {
